+++ src/client/protocolgame.h
@@ -184,7 +184,7 @@ public:
     // Wheel of Destiny
     void sendOpenWheel(uint32_t playerId);
     void sendApplyWheelPoints(const std::vector<uint16_t>& slotPoints,uint16_t greenGem,uint16_t redGem,uint16_t acquaGem,uint16_t purpleGem);
-    void sendWheelGemAction(const uint8_t actionType, const uint8_t param, const uint8_t pos);
+    void sendWheelGemAction(const uint8_t actionType, const uint16_t param, const uint8_t pos);
 
     // Weapon Proficiency
     void sendWeaponProficiencyAction(uint8_t actionType, uint16_t itemId = 0);
